refactor(client): fix tableRow
