+++ packages/client/src/components/table/table.module.css
@@ -3,7 +3,7 @@
   margin-bottom: 30px;
 }
 
-.table__title {
+.tableTitle {
   font-family: 'Rubik Bubbles', fantasy;
   font-weight: 400;
   font-size: 40px;
