+++ packages/client/src/components/tableRow/tableRow.module.css
@@ -1,4 +1,4 @@
-.table__row {
+.tableRow {
   font-family: 'Rubik Bubbles', fantasy;
   font-weight: 400;
   font-size: 24px;
